requirements.txt
tensorflow~=2.7.0
numpy~=1.20.0
pandas~=1.3.4
seaborn~=0.11.2
matplotlib~=3.5.0
Pillow~=8.4.0